pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "dolt"
version = "0.1.0"
description = "Document-native ELT Engine — 비정형 문서를 AI/RAG용 구조화 데이터로 변환"
readme = "README.md"
license = "Apache-2.0"
requires-python = ">=3.10"
authors = [{ name = "DoLT Contributors" }]
keywords = ["rag", "document", "elt", "embedding", "vector", "chunking"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: Apache Software License",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Text Processing",
]

dependencies = [
    "typer>=0.9",
    "rich>=13.0",
    "pydantic>=2.0",
    "pydantic-settings>=2.0",
    "PyMuPDF>=1.23",
    "python-docx>=1.0",
    "beautifulsoup4>=4.12",
    "markdown-it-py>=3.0",
    "tiktoken>=0.5",
    "httpx>=0.25",
    "openai>=1.0",
    "pyyaml>=6.0",
]

[project.optional-dependencies]
cohere = ["cohere>=5.0"]
local = ["sentence-transformers>=2.2"]
qdrant = ["qdrant-client>=1.7"]
pinecone = ["pinecone-client>=3.0"]
weaviate = ["weaviate-client>=4.0"]
postgres = ["psycopg2-binary>=2.9", "pgvector>=0.2"]
all = [
    "dolt[cohere,local,qdrant,pinecone,weaviate,postgres]",
]
dev = [
    "pytest>=7.0",
    "pytest-cov>=4.0",
    "pytest-asyncio>=0.21",
    "ruff>=0.1",
    "mypy>=1.0",
]

[project.scripts]
dolt = "dolt.cli.app:main"

[project.entry-points."dolt.parsers"]
[project.entry-points."dolt.metadata_plugins"]
[project.entry-points."dolt.embedding_providers"]
[project.entry-points."dolt.exporters"]

[tool.hatch.build.targets.wheel]
packages = ["src/dolt"]

[tool.ruff]
target-version = "py310"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]

[tool.mypy]
python_version = "3.10"
strict = true

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-v --cov=dolt --cov-report=term-missing"
